requirements.txt
asyncpg-lite~=0.3.1.3
aiogram~=3.7.0
python-decouple
groq
pytz